+++ packages/components/src/styles/defaults.ts
@@ -72,19 +72,22 @@ export const defaultStyles: StclandStyles = {
   button: {
     root: 'flex w-fit items-center gap-1 min-w-32 p-2.5 text-sm font-medium text-gray-800',
     primary: {
-      outlined:
-        'border border-primary-dark text-primary-main hover:border-primary-dark hover:bg-primary-range-200',
-      solid: 'bg-primary-dark hover:bg-primary-range-900 text-gray-50',
+      default: 'text-primary-surface-default',
+      hover: 'hover:bg-primary-surface-light hover:text-white',
+      pressed: 'pressed:bg-primary-surface-dark pressed:text-white',
+      disabled: 'disabled:bg-neutral-surface-disabled disabled:text-neutral-text-icon-disabled'
     },
-    secondary: {
-      outlined:
-        'border border-secondary-dark text-secondary-dark hover:border-secondary-main hover:bg-secondary-range-200',
-      solid: 'bg-secondary-dark hover:bg-secondary-range-900 text-gray-50',
+    secondary:{
+      default: 'border border-primary-surface-default text-primary-surface-default',
+      hover: 'border hover:border-primary-surface-light hover:text-primary-text-icon-default',
+      pressed: 'border pressed:border-primary-surface-light pressed:bg-primary-surface-default',
+      disabled: 'border disabled:border-neutral-text-icon-disabled disabled:text-neutral-text-icon-disabled'
     },
-    neutral: {
-      outlined:
-        'border border-gray-700 text-gray-500 hover:border-gray-600 hover:bg-gray-200',
-      solid: 'bg-gray-700 hover:bg-gray-600 text-gray-300',
+    tertiary: {
+      default: 'text-primary-surface-default',
+      hover: 'hover:text-primary-text-icon-default',
+      pressed: 'pressed:text-primary-surface-dark',
+      disabled: 'disabled:text-text-icon-disabled'
     },
     sm: 'p-2 text-1.5xs',
     md: 'p-3 text-sm',
@@ -92,15 +95,31 @@ export const defaultStyles: StclandStyles = {
     fullWidth: 'w-full',
     rounded: 'rounded-md',
     highlightOnHover: 'hover:bg-gray-600',
-    icon: 'w-1.5 h-1.5 inline',
+    leftIcon: 'w-3.5 h-3.5 inline',
+    rightIcon: 'w-3.5 h-3.5 inline',
     disabled: 'bg-gray-300 text-gray-400 hover:bg-gray-350',
-    button: 'w-full',
+    button: 'w-full'
   },
   icon: {
     root: 'p-2',
-    secondary: 'text-secondary-main',
-    primary: 'text-primary-main',
-    neutral: 'text-gray-600',
+    secondary:{
+      default: 'text-primary-surface-dark',
+      hover: 'hover:text-primary-text-icon-default',
+      pressed: 'active:text-primary-surface-dark',
+      disabled: 'disabled:bg-neutral-surface-disabled disabled:text-neutral-text-icon-disabled'
+    },
+    primary: {
+      default: 'text-primary-surface-default',
+      hover: 'hover:text-primary-surface-light',
+      pressed: 'active:text-primary-surface-dark',
+      disabled: 'disabled:bg-neutral-surface-disabled disabled:text-neutral-text-icon-disabled'
+    },
+    tertiary: {
+      default: 'text-primary-surface-light',
+      hover: 'hover:text-primary-surface-subtle',
+      pressed: 'active:text-primary-surface-dark',
+      disabled: 'disabled:bg-neutral-surface-disabled disabled:text-neutral-text-icon-disabled'
+    },
     sm: 'h-4.5 w-4.5',
     md: 'h-6 w-6',
     lg: 'h-7 w-7',
@@ -109,6 +128,8 @@ export const defaultStyles: StclandStyles = {
     highlightOnHover: 'hover:bg-gray-600',
     brightenOnHover: 'hover:brightness-125',
     muted: 'opacity-50',
+    icon: '',
+    button: ''
   },
   accordion: {
     accordion: { root: 'border rounded bg-gray-825' },
@@ -154,62 +175,95 @@ export const defaultStyles: StclandStyles = {
   },
   checkBox: {
     root: 'flex-row-reverse justify-end items-center py-4',
-    inputRoot: 'accent-sky-500 rounded-sm h-100',
+    inputRoot: {
+      default: 'w-5 h-5 rounded-sm h-100',
+      hover: '',
+      pressed: '',
+      disabled: ''
+    },
     label: 'pt-0 mb-0',
   },
   date: {
     root: 'flex flex-col w-100 gap-2',
     label: 'text-sm pt-2',
-    inputRoot:
-        'px-4 py-4 border-gray-600 border-solid rounded-lg bg-neutral-700 text-lg border text-white w-full dark:[color-scheme:dark] disabled:text-zinc-400',
+    inputRoot: {
+      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
+      hover: 'hover:border-neutral-stroke-dark',
+      pressed: 'active:text-neutral-text-icon-body',
+      disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
+    },
     error: 'text-red-300 italic text-sm',
     info: '',
   },
   email: {
     root: 'flex flex-col w-100 gap-2',
     label: 'text-sm pt-2',
-    inputRoot:
-        'px-4 py-4 border-gray-600 border-solid rounded-lg bg-neutral-700 text-lg border text-white w-full dark:[color-scheme:dark] disabled:text-zinc-400',
+    inputRoot: {
+      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
+      hover: 'hover:border-neutral-stroke-dark',
+      pressed: 'active:text-neutral-text-icon-body',
+      disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
+    },
     error: 'text-red-300 italic text-sm',
     info: '',
   },
   number: {
     root: 'flex flex-col w-100 gap-2',
     label: 'text-sm pt-2',
-    inputRoot:
-        'px-4 py-4 border-gray-600 border-solid rounded-lg bg-neutral-700 text-lg border text-white w-full dark:[color-scheme:dark] disabled:text-zinc-400',
+    inputRoot: {
+      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
+      hover: 'hover:border-neutral-stroke-dark',
+      pressed: 'active:text-neutral-text-icon-body',
+      disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
+    },
     error: 'text-red-300 italic text-sm',
     info: '',
   },
   password: {
     root: 'flex flex-col w-100 gap-2',
     label: 'text-sm pt-2',
-    inputRoot:
-        'px-4 py-4 border-gray-600 border-solid rounded-lg bg-neutral-700 text-lg border text-white w-full dark:[color-scheme:dark] disabled:text-zinc-400',
+    inputRoot: {
+      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
+      hover: 'hover:border-neutral-stroke-dark',
+      pressed: 'active:text-neutral-text-icon-body',
+      disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
+    },
     error: 'text-red-300 italic text-sm',
     info: '',
   },
   text: {
     root: 'flex flex-col w-100 gap-2',
     label: 'text-sm pt-2',
-    inputRoot:
-        'px-4 py-4 border-gray-600 border-solid rounded-lg bg-neutral-700 text-lg border text-white w-full dark:[color-scheme:dark] disabled:text-zinc-400',
+    inputRoot: {
+      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
+      hover: 'hover:border-neutral-stroke-dark',
+      pressed: 'active:text-neutral-text-icon-body',
+      disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
+    },
     error: 'text-red-300 italic text-sm',
     info: '',
   },
   textArea: {
     root: 'flex flex-col w-100 gap-2',
     label: 'text-sm pt-2',
-    inputRoot:
-        'h-14 px-4 py-4 border-gray-600 border-solid rounded-lg bg-neutral-700 text-lg border text-white w-full dark:[color-scheme:dark] disabled:text-zinc-400',
+    inputRoot: {
+      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
+      hover: 'hover:border-neutral-stroke-dark',
+      pressed: 'active:text-neutral-text-icon-body',
+      disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
+    },
     error: 'text-red-300 italic text-sm',
     info: '',
   },
   time: {
     root: 'flex flex-col w-100 gap-2',
     label: 'text-sm pt-2',
-    inputRoot:
-        'px-4 py-4 border-gray-600 border-solid rounded-lg bg-neutral-700 text-lg border text-white w-full dark:[color-scheme:dark] disabled:text-zinc-400',
+    inputRoot: {
+      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
+      hover: 'hover:border-neutral-stroke-dark',
+      pressed: 'active:text-neutral-text-icon-body',
+      disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
+    },
     error: 'text-red-300 italic text-sm',
     info: '',
   },
@@ -219,7 +273,7 @@ export const defaultStyles: StclandStyles = {
     label: '',
     radioContainer: 'flex gap-2',
     error: 'text-red-300 italic text-sm',
-    radioButton: 'accent-sky-500 h-4 w-4',
+    radioButton: 'accent-sky-500 h-4 w-4 rounded-full focus:ring-offset-2 focus:ring-2 focus:ring-primary-surface-default',
   },
   selectInput: {
     root: 'flex flex-col gap-2',
@@ -253,10 +307,10 @@ export const defaultStyles: StclandStyles = {
     rootWithoutCustomIcons: 'rounded-sm border-2 flex items-center justify-center border-gray-400',
     primary: 'bg-primary-main border-primary-main',
     secondary: 'bg-secondary-main border-secondary-main',
-    neutral: 'bg-gray-600 border-gray-600',
+    tertiary: 'bg-gray-600 border-gray-600',
     uncheckedPrimary: 'border-primary-main',
     uncheckedSecondary: 'border-secondary-main',
-    uncheckedNeutral: 'border-gray-600',
+    uncheckedTertiary: 'border-gray-600',
     sm: 'w-5 h-5',
     md: 'w-6 h-6',
     lg: 'w-8 h-8',
@@ -266,7 +320,7 @@ export const defaultStyles: StclandStyles = {
     disabled: 'bg-gray-300 border-gray-300 text-gray-400 hover:bg-gray-350',
     indeterminatePrimary: 'absolute w-3/4 h-0.5',
     indeterminateSecondary: 'absolute w-3/4 h-0.5',
-    indeterminateNeutral: 'absolute w-3/4 h-0.5',
+    indeterminateTertiary: 'absolute w-3/4 h-0.5',
     labelPrimary: 'text-primary-main',
     labelSecondary: 'text-secondary-main',
     labelNeutral: 'text-gray-700',
@@ -293,6 +347,7 @@ export const defaultStyles: StclandStyles = {
       dragging: 'opacity-50',
       content: 'flex-1',
       removeButton: 'w-6 h-6 text-gray-300 hover:text-gray-500',
+      iconContainer: ''
     },
     droppable: {
       root: '',
@@ -303,17 +358,16 @@ export const defaultStyles: StclandStyles = {
       bgColorClass: 'bg-gray-825',
       dragging: 'opacity-50',
       content: 'flex-1',
-      removeButton: 'w-6 h-6 text-gray-300 hover:text-gray-500'
+      removeButton: 'w-6 h-6 text-gray-300 hover:text-gray-500',
+      iconContainer: ''
     }
   },
   radioGroup: {
     radio: {
       root: 'flex items-center gap-2',
-      inputRoot: 'sr-only',
-      radio: 'w-4 h-4 rounded-full border-2 flex justify-center items-center border-gray-825 bg-none',
+      inputRoot: 'accent-primary-surface-default h-4 w-4 rounded-full',
       radioDisabled: 'border-gray-400 bg-gray-100 cursor-not-allowed',
       selected: 'bg-primary-main',
-      innerCircle: 'w-1.5 h-1.5 rounded-full bg-gray-825',
       text: 'text-sm text-gray-825',
       textDisabled: 'text-gray-400'
     },
@@ -326,10 +380,13 @@ export const defaultStyles: StclandStyles = {
   select: {
     root: 'relative w-64',
     label: 'block mb-2 text-sm font-medium text-gray-700',
-    button: 'flex flex-row items-center justify-between w-full px-4 py-2 text-left bg-neutral-900 border border-gray-400 text-neutral-400 rounded focus:outline-none',
-    menu: 'absolute z-10 w-full bg-zinc-700 border border-gray-300 text-neutral-400 rounded shadow-lg max-h-60 overflow-auto',
+    button: 'flex flex-row items-center justify-between w-full px-4 py-2 text-left bg-neutral-900 border border-gray-400 text-neutral-400 rounded focus:border-primary-surface-dark',
+    menu: 'absolute z-10 w-full bg-neutral-surface-1 border border-gray-300 text-neutral-text-icon-body rounded shadow-lg max-h-60 overflow-auto',
     disabled: 'disabled:bg-gray-400',
-    optionContainer: 'flex items-center p-2',
+    optionContainer: {
+      default: 'flex items-center hover:bg-neutral-surface-2 justify-between p-2 bg-primary-surface-subtle cursor-pointer',
+      selected: 'bg-primary-surface-subtle'
+    },
     listItem: {
       base: 'ml-10',
       selected: 'ml-0',
@@ -354,10 +411,10 @@ export const defaultStyles: StclandStyles = {
       rootWithoutCustomIcons: 'rounded-sm border-2 flex items-center justify-center border-gray-400',
       primary: 'bg-zinc-900 border-zinc-900 text-white',
       secondary: 'bg-yellow-400 border-yellow-600 text-zinc-900',
-      neutral: 'bg-gray-600 border-gray-600 text-white',
+      tertiary: 'bg-gray-600 border-gray-600 text-white',
       uncheckedPrimary: 'border-zinc-900',
       uncheckedSecondary: 'border-yellow-600',
-      uncheckedNeutral: 'border-gray-600',
+      uncheckedTertiary: 'border-gray-600',
       sm: 'w-5 h-5',
       md: 'w-6 h-6',
       lg: 'w-8 h-8',
@@ -367,7 +424,7 @@ export const defaultStyles: StclandStyles = {
       disabled: 'bg-gray-300 border-gray-300 text-gray-400 hover:bg-gray-350 cursor-auto',
       indeterminatePrimary: 'absolute w-3/4 h-0.5 bg-primary-main',
       indeterminateSecondary: 'absolute w-3/4 h-0.5',
-      indeterminateNeutral: 'absolute w-3/4 h-0.5',
+      indeterminateTertiary: 'absolute w-3/4 h-0.5',
       labelPrimary: 'text-zinc-900',
       labelSecondary: 'text-yellow-200',
       labelNeutral: 'text-gray-700',
@@ -375,7 +432,7 @@ export const defaultStyles: StclandStyles = {
     }
   },
   tooltip: {
-    root: 'relative inline-block',
+    root: 'relative',
     tooltipContainer: 'absolute z-50 p-3 rounded-md shadow-md text-sm w-max',
     colorClass: 'bg-zinc-600 text-white',
     arrow: 'absolute w-3.5 h-2.5 border-t-2 border-r-2 border-transparent',