+++ packages/storybook/stc.config.ts
@@ -16,35 +16,54 @@ const styles: StclandStyles = {
   button: {
     root: 'flex w-fit items-center gap-1 min-w-32 p-2.5 text-sm font-medium text-gray-800',
     primary: {
-      outlined:
-        'border border-primary-range-300 text-primary-main hover:border-primary-dark hover:bg-primary-range-200',
-      solid: 'bg-secondary-main hover:bg-primary-range-900 text-primary-main',
+      default: 'bg-yellow-900 text-white',
+      hover: 'hover:bg-green-400 hover:text-yellow-900 border-2 border-yellow-900',
+      pressed: 'active:bg-primary-surface-dark active:text-yellow-700',
+      disabled: 'disabled:bg-neutral-surface-disabled disabled:text-neutral-text-icon-disabled'
     },
-    secondary: {
-      outlined:
-        'border border-secondary-main text-secondary-main hover:border-secondary-dark hover:bg-secondary-range-200',
-      solid: 'bg-secondary-main hover:bg-secondary-range-900 text-gray-50',
+    secondary:{
+      default: 'border-2 border-primary-surface-subtle text-primary-surface-default',
+      hover: 'border-2 hover:border-primary-surface-subtle hover:text-primary-text-icon-default',
+      pressed: 'border-2 active:border-primary-surface-light active:bg-primary-surface-default active:text-primary-text-icon-default',
+      disabled: 'border-2 disabled:border-neutral-text-icon-disabled disabled:text-neutral-text-icon-disabled'
     },
-    neutral: {
-      outlined:
-        'border border-gray-700 text-gray-500 hover:border-gray-600 hover:bg-gray-200',
-      solid: 'bg-gray-700 hover:bg-gray-600 text-gray-300',
+    tertiary: {
+      default: 'text-primary-surface-default',
+      hover: 'hover:text-primary-text-icon-default',
+      pressed: 'active:text-primary-surface-dark',
+      disabled: 'disabled:text-text-icon-disabled'
     },
     sm: 'p-2 text-1.5xs',
     md: 'p-3 text-sm',
     lg: 'p-3 text-md',
-    // fullWidth: 'w-full',
-    // rounded: 'rounded-md',
+    fullWidth: 'w-full',
+    rounded: 'rounded-md',
     highlightOnHover: 'hover:bg-gray-600',
-    icon: 'w-3 h-3 inline',
-    // disabled:  'bg-gray-300 text-gray-400 hover:bg-gray-350',
-    // button: 'w-full'
+    leftIcon: 'w-5 h-5 inline',
+    rightIcon: 'w-5 h-5 inline',
+    disabled: 'bg-gray-300 text-gray-400 hover:bg-gray-350',
+    button: 'w-full'
   },
   icon: {
     root: 'p-2',
-    secondary: 'text-gray-800',
-    primary: 'text-gray-100',
-    neutral: 'text-gray-600',
+    secondary:{
+      default: 'text-pink-700',
+      hover: 'hover:text-pink-500',
+      pressed: 'active:text-pink-900',
+      disabled: 'disabled:bg-neutral-surface-disabled disabled:text-neutral-text-icon-disabled'
+    },
+    primary: {
+      default: 'text-cyan-700',
+      hover: 'hover:text-cyan-300',
+      pressed: 'active:text-cyan-900',
+      disabled: 'disabled:bg-neutral-surface-disabled disabled:text-neutral-text-icon-disabled'
+    },
+    tertiary: {
+      default: 'text-gray-700',
+      hover: 'hover:text-gray-900',
+      pressed: 'active:text-text-500',
+      disabled: 'disabled:bg-neutral-surface-disabled disabled:text-neutral-text-icon-disabled'
+    },
     sm: 'h-4.5 w-4.5',
     md: 'h-6 w-6',
     lg: 'h-7 w-7',
@@ -53,6 +72,8 @@ const styles: StclandStyles = {
     highlightOnHover: 'hover:bg-gray-600',
     brightenOnHover: 'hover:brightness-125',
     muted: 'opacity-50',
+    icon: 'text-states-error-dark',
+    button: ''
   },
   accordion: {
     accordion: { root: 'border rounded bg-gray-700' },
@@ -71,39 +92,70 @@ const styles: StclandStyles = {
   email: {
     root: 'flex  w-100 gap-2 ',
     label: 'text-md text-sky-500',
-    inputRoot: 'px-2 py-2 bg-slate-300 border-sky-200 bg-white text-slate-800',
-  },
+    inputRoot: {
+      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
+      hover: 'hover:border-neutral-stroke-dark',
+      pressed: 'active:text-neutral-text-icon-body',
+      disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
+    },  },
   number: {
     root: 'flex  w-100 gap-2 ',
     label: 'text-md text-sky-500',
-    inputRoot: 'px-2 py-2 bg-slate-300 border-sky-200 bg-white text-slate-800',
+    inputRoot: {
+      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
+      hover: 'hover:border-neutral-stroke-dark',
+      pressed: 'active:text-neutral-text-icon-body',
+      disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
+    },
   },
   password: {
     root: 'flex  w-100 gap-2 ',
     label: 'text-md text-sky-500',
-    inputRoot: 'px-2 py-2 bg-slate-300 border-sky-200 bg-white text-slate-800',
+    inputRoot: {
+      default: 'text-base px-4 py-4 border-sky-200 border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
+      hover: 'hover:border-neutral-stroke-dark',
+      pressed: 'active:text-neutral-text-icon-body',
+      disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
+    },
   },
   date: {
     root: 'flex  w-100 gap-2 ',
     label: 'text-md text-sky-500',
-    inputRoot:
-      'px-2 py-2 bg-slate-300 border-sky-200 bg-white text-slate-800 dark:[color-scheme:light]',
+    inputRoot: {
+      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
+      hover: 'hover:border-neutral-stroke-dark',
+      pressed: 'active:text-neutral-text-icon-body',
+      disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
+    },
   },
   text: {
     root: 'flex  w-100 gap-2 ',
     label: 'text-md text-sky-500',
-    inputRoot: 'px-2 py-2 bg-slate-300 border-sky-200 bg-white text-slate-800',
+    inputRoot: {
+      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
+      hover: 'hover:border-neutral-stroke-dark',
+      pressed: 'active:text-neutral-text-icon-body',
+      disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
+    },
   },
   textArea: {
     root: 'flex  w-100 gap-2 ',
     label: 'text-md text-sky-500',
-    inputRoot: 'px-2 py-2 bg-slate-300 border-sky-200 bg-white text-slate-800',
-  },
+    inputRoot: {
+      default: 'text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
+      hover: 'hover:border-neutral-stroke-dark',
+      pressed: 'active:text-neutral-text-icon-body',
+      disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
+    },  },
   time: {
     root: 'flex  w-100 gap-2 ',
     label: 'text-md text-sky-500',
-    inputRoot:
-      'px-2 py-2 bg-slate-300 border-sky-200 bg-white text-slate-800 dark:[color-scheme:light]',
+    inputRoot: {
+      default: 'p-2 text-base px-4 py-4 border-neutral-stroke-default border-2 rounded-lg bg-neutral-surface-2 text-neutral-text-icon-body border w-full',
+      hover: 'hover:border-neutral-stroke-dark',
+      pressed: 'active:text-neutral-text-icon-body',
+      disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
+    },
   },
   fieldSet: {
     label: 'font-semibold text-sky-700 text-lg',
@@ -111,8 +163,12 @@ const styles: StclandStyles = {
   },
   checkBox: {
     root: 'flex-row-reverse justify-end items-center py-4',
-    inputRoot:
-      'accent-sky-500 rounded-sm h-100 bg-white border-white dark:[color-scheme:light]',
+    inputRoot: {
+      default: 'w-5 h-5 text-base px-4 py-4 border-pink-600 border-2 rounded-lg  text-pink-900 border',
+      hover: 'hover:border-neutral-stroke-dark',
+      pressed: 'active:text-pink-600',
+      disabled: 'disabled:text-neutral-stroke-default disabled:bg-neutral-surface-disabled'
+    },
     label: 'pt-0 mb-0 text-sky-500 text-md',
   },
   selectInput: {
@@ -164,26 +220,29 @@ const styles: StclandStyles = {
     },
   },
   checkbox: {
-    container: 'flex flex-row gap-2',
-    root: '',
-    rootWithoutCustomIcons: 'rounded-sm border-2 flex items-center justify-center border-gray-800',
-    primary: 'bg-pink-400 border-pink-400',
-    secondary: 'bg-purple-400 border-purple-400',
-    neutral: 'bg-gray-600 border-gray-600',
-    uncheckedPrimary: 'border-pink-400',
-    uncheckedSecondary: 'border-purple-400',
-    uncheckedNeutral: 'border-gray-600',
+    container: 'flex flex-row',
+    root: 'cursor-pointer',
+    rootWithoutCustomIcons: 'rounded-sm border-2 flex items-center justify-center border-gray-400',
+    primary: 'bg-primary-main border-primary-main',
+    secondary: 'bg-secondary-main border-secondary-main',
+    tertiary: 'bg-gray-600 border-gray-600',
+    uncheckedPrimary: 'border-primary-main',
+    uncheckedSecondary: 'border-secondary-main',
+    uncheckedTertiary: 'border-gray-600',
     sm: 'w-5 h-5',
     md: 'w-6 h-6',
     lg: 'w-8 h-8',
     smChecked: 'text-xs',
     mdChecked: 'text-md',
     lgChecked: 'text-lg',
-    disabled: 'bg-gray-800 border-gray-800 text-gray-600 hover:bg-gray-750',
-    indeterminatePrimary: 'absolute w-3/4 h-0.5 bg-gray-200',
+    disabled: 'bg-gray-300 border-gray-300 text-gray-400 hover:bg-gray-350',
+    indeterminatePrimary: 'absolute w-3/4 h-0.5',
     indeterminateSecondary: 'absolute w-3/4 h-0.5',
-    indeterminateNeutral: 'absolute w-3/4 h-0.5',
-    labelPrimary: 'text-pink-400'
+    indeterminateTertiary: 'absolute w-3/4 h-0.5',
+    labelPrimary: 'text-primary-main',
+    labelSecondary: 'text-secondary-main',
+    labelNeutral: 'text-gray-700',
+    labelCustomIcon: 'text-gray-925'
   },
   modal: {
     root: 'fixed inset-0 z-50 flex items-center justify-center',
@@ -201,11 +260,12 @@ const styles: StclandStyles = {
       grid: 'grid grid-cols-3 gap-4 p-4 bg-gray-200 rounded-md'
     },
     sortableItem: {
-      root: 'relative p-2 rounded text-pink-900 transition-opacity flex items-center justify-between',
-      bgColorClass: 'bg-pink-200',
+      root: 'relative p-1 rounded text-zinc-100 transition-opacity flex items-center justify-between',
+      bgColorClass: 'bg-indigo-600',
       dragging: 'opacity-50',
       content: 'flex-1',
-      removeButton: 'w-6 h-6 text-gray-300 hover:text-gray-500',
+      removeButton: 'w-6 h-6 text-zinc-100 hover:text-gray-500',
+      iconContainer: ''
     },
     droppable: {
       root: '',
@@ -213,35 +273,32 @@ const styles: StclandStyles = {
     },
     draggable: {
       root: 'cursor-grab relative p-2 rounded text-primary-main transition-opacity flex items-center justify-between',
-      bgColorClass: 'bg-pink-200',
+      bgColorClass: ' bg-pink-200',
       dragging: 'opacity-50',
       content: 'flex-1',
-      removeButton: 'w-6 h-6 text-gray-300 hover:text-gray-500'
+      removeButton: 'w-6 h-6 text-zinc-900 hover:text-gray-500',
+      iconContainer: ''
     }
   },
   radioGroup: {
     radio: {
       root: 'flex items-center gap-2',
-      inputRoot: 'sr-only',
-      radio: 'w-4 h-4 rounded-full border-2 flex justify-center items-center border-primary-main bg-none',
+      inputRoot: 'accent-primary-surface-default h-4 w-4 rounded-full',
       radioDisabled: 'border-gray-400 bg-gray-100 cursor-not-allowed',
       selected: 'bg-primary-main',
-      innerCircle: 'w-1.5 h-1.5 rounded-full bg-gray-800',
-      text: 'text-sm text-gray-200',
+      text: 'text-sm text-primary-surface-dark',
       textDisabled: 'text-gray-400'
     },
     radioGroup: {
-      root: 'flex bg-gray-800 p-4 rounded-lg',
+      root: 'flex bg-primary-surface-subtle p-4 rounded-lg',
       vertical: 'flex-col space-y-2',
       horizontal: 'flex-row space-x-4'
     },
     isolatedRadio: {
-      root: 'flex items-center gap-2',
-      inputRoot: 'sr-only',
-      radio: 'w-4 h-4 rounded-full border-2 flex justify-center items-center border-gray-800 bg-none',
+      root: 'flex items-center gap-2 rounded-full',
+      inputRoot: 'accent-pink-700 h-4 w-4 rounded-full',
       radioDisabled: 'border-gray-400 bg-gray-100 cursor-not-allowed',
       selected: 'bg-pink-500',
-      innerCircle: 'w-1.5 h-1.5 rounded-full bg-gray-800',
       text: 'text-sm text-pink-500',
       textDisabled: 'text-gray-200'
     }
@@ -249,15 +306,17 @@ const styles: StclandStyles = {
   select: {
     root: 'relative w-64',
     label: 'block mb-2 text-sm font-medium text-pink-700',
-    button: 'flex flex-row items-center justify-between w-full px-4 py-2 text-left bg-pink-900 border border-gray-400 text-neutral-200 rounded focus:outline-none',
-    menu: 'absolute z-10 w-full bg-pink-600 border border-gray-300 text-neutral-700 rounded shadow-lg max-h-60 overflow-auto',
-    disabled: 'disabled:bg-gray-400',
-    optionContainer: 'flex items-center p-2',
-    listItem: {
+    button: 'flex flex-row items-center justify-between w-full px-4 py-2 text-left bg-pink-900 border border-gray-400 text-neutral-200 rounded focus:primary-surface-dark',
+    menu: 'absolute z-10 w-full bg-neutral-surface-1 border border-gray-300 text-neutral-text-icon-body rounded shadow-lg max-h-60 overflow-auto',
+    disabled: 'text-neutral-text-icon-disabled',
+    optionContainer: {
+      default: 'flex items-center hover:bg-pink-400 justify-between p-2 bg-pink-900 cursor-pointer',
+      selected: 'bg-pink-200'
+    },     listItem: {
       base: 'ml-10',
       selected: 'ml-0',
     },
-    selectedDefaultIcon: 'text-green-300'
+    selectedDefaultIcon: 'text-pink-900'
   },
   divider: {
     root: 'border-0',
@@ -275,12 +334,12 @@ const styles: StclandStyles = {
       container: 'flex flex-row items-center gap-2',
       root: 'cursor-pointer',
       rootWithoutCustomIcons: 'rounded-sm border-2 flex items-center justify-center border-gray-400',
-      primary: 'bg-purple-500 border-purple-500 text-white',
-      secondary: 'bg-green-500 border-green-500 text-white',
-      neutral: 'bg-gray-600 border-gray-600 text-white',
-      uncheckedPrimary: 'border-purple-500',
-      uncheckedSecondary: 'border-green-500',
-      uncheckedNeutral: 'border-gray-600',
+      primary: 'bg-zinc-900 border-zinc-900 text-white',
+      secondary: 'bg-yellow-400 border-yellow-600 text-zinc-900',
+      tertiary: 'bg-gray-600 border-gray-600 text-white',
+      uncheckedPrimary: 'border-zinc-900',
+      uncheckedSecondary: 'border-yellow-600',
+      uncheckedTertiary: 'border-gray-600',
       sm: 'w-5 h-5',
       md: 'w-6 h-6',
       lg: 'w-8 h-8',
@@ -290,9 +349,9 @@ const styles: StclandStyles = {
       disabled: 'bg-gray-300 border-gray-300 text-gray-400 hover:bg-gray-350 cursor-auto',
       indeterminatePrimary: 'absolute w-3/4 h-0.5 bg-primary-main',
       indeterminateSecondary: 'absolute w-3/4 h-0.5',
-      indeterminateNeutral: 'absolute w-3/4 h-0.5',
-      labelPrimary: 'text-purple-500',
-      labelSecondary: 'text-green-500',
+      indeterminateTertiary: 'absolute w-3/4 h-0.5',
+      labelPrimary: 'text-zinc-900',
+      labelSecondary: 'text-yellow-200',
       labelNeutral: 'text-gray-700',
       labelCustomIcon: 'text-gray-925'
     }